backend/config.py
import os
from dotenv import load_dotenv

load_dotenv()  # Load variables from .env

class Config:
    SECRET_KEY = os.getenv("SECRET_KEY", "dev-fallback-key")
    OPENROUTER_API_KEY = os.getenv("OPENROUTER_API_KEY")
    OPENROUTER_BASE_URL = "https://openrouter.ai/api/v1"
    # Use a working FREE model (as of Nov 2025)
    AI_MODEL = "google/gemma-2-2b-it:free"