backend/requirements.txt
Flask==3.0.3
openai==1.40.0
python-dotenv==1.0.1
Flask-Cors==5.0.0